requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.3.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudinary==1.41.0
colorama==0.4.6
cryptography==44.0.2
cssselect2==0.7.0
distlib==0.3.9
Django==5.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-filter==24.3
django-user-agents==0.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.8
filelock==3.16.1
html5lib==1.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.1
Markdown==3.7
matplotlip==0.2
mysql-connector-python==9.2.0
mysqlclient==2.2.6
ngrok==1.4.0
oscrypto==1.3.0
packaging==24.2
pdfkit==1.0.0
pillow==11.0.0
pipenv==2024.4.0
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
pycparser==2.22
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.10.1
PyMySQL==1.1.1
pypdf==5.3.0
pytest==8.3.4
python-bidi==0.6.6
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
referencing==0.35.1
reportlab==4.3.1
requests==2.32.3
rpds-py==0.22.3
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzdata==2024.2
tzlocal==5.3
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
user-agents==2.2.0
virtualenv==20.28.0
webencodings==0.5.1
wkhtmltopdf==0.2
xhtml2pdf==0.2.17
